Database Management/week 03/Module 11: Python & MySQL/11-2 Execute MySQL Query in Python/02-create-db.py
import mysql.connector

mydbconnection=mysql.connector.connect(
  host="localhost",
  user='root',
  passwd='root'
)

# create database
db_name='python_test_db'
mycoursor = mydbconnection.cursor()

sql_query= "CREATE DATABASE " + db_name

mycoursor.execute(sql_query)


